[Generator] Move public headers in to interface directory
